\name{spatstat.univar-internal}
\title{Internal Functions}
\alias{as.breakpts}
\alias{breakpts}
\alias{breakpts.from.r}
\alias{check.hist.lengths}
\alias{make.even.breaks}
\alias{match.kernel}
\alias{print.ewcdf}
\alias{StieltjesCalc}
\alias{StieltjesCalc.stepfun}
\description{
  Undocumented Functions
}
\usage{
as.breakpts(\dots)
breakpts(val, maxi, even, npos, step)
breakpts.from.r(r)
check.hist.lengths(hist, breaks)
make.even.breaks(bmax, npos, bstep)
match.kernel(kernel)
\method{print}{ewcdf}(x, digits, \dots)
StieltjesCalc(M, f, \dots)
\method{StieltjesCalc}{stepfun}(M, f, \dots)
}
\details{
  These functions are not documented.
}
\keyword{internal}
